refactor: extract Amazon.getProductTitle
