+++ android/app/src/main/java/com/dismoi/scout/accessibility/BackgroundService.kt
@@ -16,6 +16,8 @@ import android.view.accessibility.AccessibilityNodeInfo
 import androidx.annotation.RequiresApi
 import com.dismoi.scout.accessibility.BackgroundModule.Companion.sendEventFromAccessibilityServicePermission
 import com.dismoi.scout.accessibility.browser.Chrome
+import com.dismoi.scout.browser.Amazon
+import com.dismoi.scout.browser.Helpers
 import com.facebook.react.HeadlessJsTaskService
 
 class BackgroundService : AccessibilityService() {
@@ -23,8 +25,6 @@ class BackgroundService : AccessibilityService() {
   private var _eventTime: String? = ""
   private var _packageName: String? = ""
   val chrome: Chrome = Chrome()
-
-  private val NOTIFICATION_TIMEOUT: Long = 500
 
   private val TAG = "Accessibility"
 
@@ -71,28 +71,10 @@ class BackgroundService : AccessibilityService() {
 
     val info = serviceInfo
     
-    // Set the type of events that this service wants to listen to. Others
-    // won't be passed to this service
-    /* 
-      Represents the event of changing the content of a window and more specifically 
-      the sub-tree rooted at the event's source
-    */
     info.eventTypes = AccessibilityEvent.TYPE_WINDOW_STATE_CHANGED or AccessibilityEvent.TYPE_VIEW_ACCESSIBILITY_FOCUSED or AccessibilityEvent.TYPE_WINDOW_CONTENT_CHANGED
-
-    //AccessibilityEvent.TYPE_WINDOWS_CHANGED or
-    /*
-      info.packageNames is not set because we want to receive event from
-      all packages
-     */
-
+    // info.packageNames is not set because we want to receive event from all packages
     info.feedbackType = AccessibilityServiceInfo.FEEDBACK_VISUAL
     info.flags = AccessibilityServiceInfo.FLAG_REPORT_VIEW_IDS or AccessibilityServiceInfo.FLAG_INCLUDE_NOT_IMPORTANT_VIEWS
-
-    /* 
-      The minimal period in milliseconds between two accessibility events of 
-      the same type are sent to this service
-    */
-    info.notificationTimeout = NOTIFICATION_TIMEOUT
 
     this.serviceInfo = info
   }
@@ -119,117 +101,6 @@ class BackgroundService : AccessibilityService() {
     return "com.android.launcher3" == packageName
   }
 
-  @RequiresApi(Build.VERSION_CODES.O)
-  private fun findByClassName(node: AccessibilityNodeInfo, className: String, level: Int = 0): AccessibilityNodeInfo? {
-    node.refresh()
-    val count = node.childCount
-    for (i in 0 until count) {
-      val child = node.getChild(i)
-      if (child != null) {
-        if (child.className.toString() == className) {
-          return child
-        }
-        val foundInChild = findByClassName(child, className, level + 1)
-        if (foundInChild != null) return foundInChild
-      }
-    }
-    return null
-  }
-
-  @RequiresApi(Build.VERSION_CODES.P)
-  private fun findHeading(node: AccessibilityNodeInfo, level: Int = 0): AccessibilityNodeInfo? {
-    node.refresh()
-    val count = node.childCount
-    for (i in 0 until count) {
-      val child = node.getChild(i)
-      if (child != null) {
-        if (child?.isHeading) {
-          return child
-        }
-        val foundInChild = findHeading(child, level + 1)
-        if (foundInChild != null) return foundInChild
-      }
-    }
-    return null
-  }
-
-
-  @RequiresApi(Build.VERSION_CODES.O)
-  private fun findById(node: AccessibilityNodeInfo, id: String, level: Int = 0): AccessibilityNodeInfo? {
-    val count = node.childCount
-    for (i in 0 until count) {
-      val child = node.getChild(i)
-
-      if (child != null) {
-        if (child?.viewIdResourceName?.toString() == id) {
-          return child
-        }
-        val foundInChild = findById(child, id, level + 1)
-        if (foundInChild != null) return foundInChild
-      }
-    }
-    return null
-  }
-
-
-  @RequiresApi(Build.VERSION_CODES.O)
-  private fun findTexts(node: AccessibilityNodeInfo, level: Int = 0): String {
-    val count = node.childCount
-    var texts = ""
-    for (i in 0 until count) {
-      val child = node.getChild(i)
-
-      texts += child.text?.toString() ?: child.contentDescription?.toString() ?: ""
-      texts += "\n" + findTexts(child, level + 1) + "\n"
-    }
-    return texts
-  }
-
-  @RequiresApi(Build.VERSION_CODES.O)
-  private fun findWebview(node: AccessibilityNodeInfo): AccessibilityNodeInfo? {
-    return findByClassName(node, "android.webkit.WebView")
-  }
-
-  @RequiresApi(30)
-  private fun logHierarchy(node: AccessibilityNodeInfo, level: Int = 0) {
-    node.refresh()
-    val id = node.viewIdResourceName
-    val text = node.text?.toString()
-    val content = node.contentDescription?.toString()
-    val avExtras = node.availableExtraData?.joinToString(", ")
-    val count = node.childCount
-
-    val extras = node.extras
-    val extrasList = mutableListOf<String>()
-    for (key in extras.keySet()) {
-      extrasList.add("$key: ${extras.get(key)}")
-    }
-    val allExtras = extrasList.joinToString(" / ")
-
-    Log.d(TAG, "${"  ".repeat(level)} ($level) " +
-      "className: ${node.className}, " +
-      "id: ${id ?: "NO ID"}, " +
-      "text: $text, " +
-      "content: $content, " +
-      "extras: $allExtras, " +
-      "avExtras: $avExtras, " +
-      "hint: ${node.hintText}, " +
-      "heading: ${node.isHeading}, " +
-      "inputType: ${node.inputType}, " +
-      "state: ${node.stateDescription}"
-    )
-
-    for (i in 0 until count)  {
-      val child = node.getChild(i)
-      if (child != null) {
-        logHierarchy(child, level + 1)
-      }
-    }
-  }
-
-  private fun getTextAndContent(node: AccessibilityNodeInfo?): String? {
-    return if (node != null)  "${node.text?.toString()} / ${node.contentDescription?.toString()}" else null
-  }
 
   /*
     This method is called back by the system when it detects an 
@@ -242,23 +113,20 @@ class BackgroundService : AccessibilityService() {
 
     val root = rootInActiveWindow
 
-    if (root == null) {
+    if (
+      root != null
+      && (root.packageName?.toString() == "com.android.chrome" || root.packageName?.toString() == "org.mozilla.firefox" )
+      && Build.VERSION.SDK_INT >= Build.VERSION_CODES.O
+    ) {
+      Log.d(TAG, "Active window packageName : ${root.packageName}, className: ${root.className}")
+    } else {
+      Log.d(TAG, "Unknown window packageName : ${root?.packageName}, className: ${root?.className}")
       return
     }
 
-    if (root.packageName?.toString() == "com.android.chrome" && Build.VERSION.SDK_INT >= Build.VERSION_CODES.O && root != null) {
-      Log.d(TAG, "Active window packageName : ${root.packageName}, className: ${root.className}")
-    }
-
-    // Temporary demo code specific to Amazon
-    val webview = findWebview(root)
+    val webview = Helpers.findWebview(root)
     if (webview != null) {
-      val titleExpanderContent = findById(webview, "titleExpanderContent")
-      if (titleExpanderContent != null) {
-        val titleView = findHeading(titleExpanderContent)
-        val title = titleView?.text ?: titleView?.contentDescription
-        Log.d(TAG, "Found Amazon page title : $title")
-      }
+      Amazon.getProductTitle(webview)
     }
 
     if (overlayIsActivated(applicationContext)) {applicationContext
